requirements_pt1.txt
absl-py==2.3.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cmake==3.30.3
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dlib==19.24.6
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.4
fsspec==2025.5.1
gdown==5.2.0
huggingface-hub==0.33.0
idna==3.10
importlib-metadata==8.4.0
importlib-resources==6.4.5
imutils==0.5.4
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.7
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-widgets==3.0.15
kiwisolver==1.4.7
libreface==0.1.1
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mediapipe==0.10.5
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.3.0
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
packaging==25.0
pandas==2.0.3
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.8
prompt-toolkit==3.0.51
protobuf==3.20.3
psutil==7.0.0
pure-eval==0.2.3
pycparser==2.22
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.13.1
six==1.17.0
sounddevice==0.5.2
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
timm==1.0.9
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.1
tqdm==4.66.5
traitlets==5.14.3
typing-extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.14
zipp==3.23.0
